requirements.txt
requests
psycopg2-binary
python-dotenv
feedparser
beautifulsoup4
